requirements.txt
google-api-python-client==2.90.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
requests-file==1.5.1
python-dotenv==1.0.0
pyzipper==0.3.6
